# Conflicts:
#	.gitignore
#	CMakeLists.txt
#	README.md
#	doc/mainpage.md
#	images/logo/logo_svg-01.svg
#	src/DiricheletBoundaryLib.cpp
#	src/DyakonovShurBoundaryLib.cpp
#	src/ElectricLib.cpp
#	src/Fluid1DLib.cpp
#	src/Fluid2DLib.cpp
#	src/GrapheneFluid1DLib.cpp
#	src/GrapheneFluid2DLib.cpp
#	src/RobinBoundaryLib.cpp
#	src/SetUpParametersLib.cpp
#	src/TETHYS_1D_Main_v134.cpp
#	src/TETHYS_2D_ElectronicAnalysis.cpp
#	src/TETHYS_2D_Main_v250.cpp
#	src/TethysBaseLib.cpp
#	src/includes/BoundaryLib.h
#	src/includes/DiricheletBoundaryLib.h
#	src/includes/DyakonovShurBoundaryLib.h
#	src/includes/ElectricLib.h
#	src/includes/Fluid1DLib.h
#	src/includes/Fluid2DLib.h
#	src/includes/GrapheneFluid1DLib.h
#	src/includes/GrapheneFluid2DLib.h
#	src/includes/Grid2DLib.h
#	src/includes/RobinBoundaryLib.h
#	src/includes/SetUpParametersLib.h

+++ src/ElectricLib.cpp
@@ -1,6 +1,6 @@
 #include "includes/ElectricLib.h"
-#include "GrapheneFluid2DLib.h"
-#include "GrapheneFluid1DLib.h"
+#include "includes/GrapheneFluid2DLib.h"
+#include "includes/GrapheneFluid1DLib.h"
 
 #ifndef MAT_PI
 #    define MAT_PI 3.14159265358979323846
@@ -56,7 +56,7 @@ void ElectroAnalysis::ComputeElectroBase(float t, const GrapheneFluid2D& graphen
 
 void ElectroAnalysis::ComputeElectroDerived() {
 	float dt;
-	dt=TmpArr.back()/DipX.size();
+	dt=TmpArr.back()/static_cast<float>(DipX.size());
 	EngCap.resize(DipX.size());
 	PowCap.resize(DipX.size());
 	transform(NetQ.begin(), NetQ.end(), EngCap.begin(), [](const float &c){ return 0.5f*c*c; });
